se termino con la lógica de agregar a favoritos, agregue la ventana canvas en todas las paginas y el respectivo archivo javascript
