+++ pages/detalleJuego.html
@@ -230,20 +230,7 @@
     </div>
     <div class="offcanvas-body">
       <h6>Estos son tus juegos favoritos</h6>
-      <div class="card card-customized mt-3">
-        <div class="row g-0 justify-content-center align-items-center">
-          <div class="col-md-4">
-            <img src="https://th.bing.com/th/id/R.e94b9c3ff8c3df6575fbcd191de78996?rik=iDbm%2fMSi6cnJAQ&riu=http%3a%2f%2fjuegos.es%2fwp-content%2fuploads%2fGTAVportada.jpg&ehk=6vWapvXGNSyhWSfKDpe37%2b%2bh0zBJprUsmd2%2bw7G%2bm7I%3d&risl=&pid=ImgRaw&r=0" alt="Imagen del juego" class="img-fluid">
-          </div>
-          <div class="col-md-8">
-            <div class="card-body text-end">
-              <button type="button" class="btn-close btn-close-white position-absolute top-0 end-0" aria-label="Eliminar"></button>
-              <h5 class="card-title text-light mb-4">Grand Theft Auto V</h5>
-              <a href="#" class="btn-customized">Ver detalles</a>
-            </div>
-          </div>
-        </div>
-      </div>
+      
     </div>
   </div>
 
@@ -261,7 +248,7 @@
     integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
     crossorigin="anonymous"></script>
   <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-  <script src="../js/favourites.js" type="module"></script>
+  <script src="../js/addFavourites.js" type="module"></script>
   <script src="../js/masInformacion.js" type="module"></script>
   <script src="../js/resenias.js" type="module"></script>
 </body>
