+++ pages/juegos.html
@@ -102,14 +102,9 @@
                 <a class="nav-link" href="nosotros.html">Sobre Nosotros</a>
               </li>
               <li class="nav-item">
-                <a
-                  class="nav-link"
-                  data-bs-toggle="offcanvas"
-                  href="#offcanvasExample"
-                  role="button"
-                  aria-controls="offcanvasExample"
-                  ><i class="bi bi-heart"></i
-                ></a>
+                <button class="nav-link" id="btnFavoritos">
+                  <i class="bi bi-heart"></i>
+                </button>
               </li>
               <li class="nav-item">
                 <ul class="navbar-nav">
@@ -147,12 +142,12 @@
     <main class="bg-main">
       <section class="my-5">
         <div class="container-title">
-          <h1 class="text-light fw-bold display-3 text-center">Todos los juegos</h1>
+          <h1 class="text-light fw-bold display-3 text-center">
+            Todos los juegos
+          </h1>
         </div>
         <div class="container my-5">
-          <div
-            class="row justify-content-between align-items-center gap-3"
-          >
+          <div class="row justify-content-between align-items-center gap-3">
             <div class="col-md-12 col-lg-6">
               <h1 class="text-light fs-2 fw-bold text-center text-lg-start">
                 JUEGOS DISPONIBLES
@@ -175,23 +170,31 @@
                 <button class="btn-customized-2 btn-genero mb-3">Acción</button>
               </div>
               <div class="col-6 col-sm-4 col-lg-2">
-                <button class="btn-customized-2 btn-genero mb-3">Fantasia</button>
+                <button class="btn-customized-2 btn-genero mb-3">
+                  Fantasia
+                </button>
               </div>
               <div class="col-6 col-sm-4 col-lg-2">
-                <button class="btn-customized-2 btn-genero mb-3">Aventura</button>
+                <button class="btn-customized-2 btn-genero mb-3">
+                  Aventura
+                </button>
               </div>
               <div class="col-6 col-sm-4 col-lg-2">
                 <button class="btn-customized-2 btn-genero mb-3">Terror</button>
               </div>
               <div class="col-6 col-sm-4 col-lg-2">
-                <button class="btn-customized-2 btn-genero mb-3">Animado</button>
+                <button class="btn-customized-2 btn-genero mb-3">
+                  Animado
+                </button>
               </div>
               <div class="col-6 col-sm-4 col-lg-2">
-                <button class="btn-customized-2 btn-genero mb-3">Deporte</button>
+                <button class="btn-customized-2 btn-genero mb-3">
+                  Deporte
+                </button>
               </div>
             </div>
           </div>
-          <hr class="text-info">
+          <hr class="text-info" />
           <div class="row my-4 justify-content-center px-1 px-sm-0">
             <div class="col-sm-6 col-md-5 col-lg-3 mb-5">
               <div class="card card-customized pb-3 h-100">
@@ -340,6 +343,28 @@
       </section>
     </main>
 
+    <div
+      class="offcanvas offcanvas-end bg-main text-light"
+      tabindex="-1"
+      id="offcanvasFavoritos"
+      aria-labelledby="offcanvasFavoritosLabel"
+    >
+      <div class="offcanvas-header">
+        <h5 class="offcanvas-title" id="offcanvasFavoritosLabel">
+          Lista de Favoritos
+        </h5>
+        <button
+          type="button"
+          class="btn-close btn-close-white"
+          data-bs-dismiss="offcanvas"
+          aria-label="Close"
+        ></button>
+      </div>
+      <div class="offcanvas-body">
+        <h6>Estos son tus juegos favoritos</h6>
+      </div>
+    </div>
+
     <footer>
       <img
         src="../img/logo.png"
@@ -361,5 +386,6 @@
       integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
       crossorigin="anonymous"
     ></script>
+    <script type="module" src="../js/favourites.js"></script>
   </body>
 </html>
